+++ Alexandria.vcxproj.filters
@@ -18,9 +18,6 @@
     <ClInclude Include="src\position.h">
       <Filter>File di intestazione</Filter>
     </ClInclude>
-    <ClInclude Include="src\hashkey.h">
-      <Filter>File di intestazione</Filter>
-    </ClInclude>
     <ClInclude Include="src\init.h">
       <Filter>File di intestazione</Filter>
     </ClInclude>
@@ -99,9 +96,6 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\position.cpp">
-      <Filter>File di origine</Filter>
-    </ClCompile>
-    <ClCompile Include="src\hashkey.cpp">
       <Filter>File di origine</Filter>
     </ClCompile>
     <ClCompile Include="src\main.cpp">
