--- src/hashkey.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "types.h"
-
-struct Position;
-
-[[nodiscard]] Bitboard GeneratePosKey(const Position* pos);
