+++ src/position.h
@@ -186,6 +186,8 @@ constexpr char ascii_pieces[13] = "PNBRQKpnbrqk";
 // NNUE
 extern NNUE nnue;
 
+// Generates zobrist key from scratch
+[[nodiscard]] Bitboard GeneratePosKey(const Position* pos);
 // parse FEN string
 void ParseFen(const std::string& command, Position* pos);
 // Get fen string from board
